campus_market_env/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-campus_market_env"
version = "0.1.0"
description = "Campus Market reinforcement-learning environment package"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.115,<1.0",
  "numpy>=1.26",
  "openenv-core>=0.2.3",
  "pydantic>=2.8,<3.0",
  "uvicorn>=0.30,<1.0",
]

[project.optional-dependencies]
gym = [
  "gymnasium>=0.29,<1.0",
]

[project.scripts]
server = "campus_market_env.server.app:main"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["campus_market_env*"]

[tool.setuptools.package-data]
campus_market_env = ["openenv.yaml", "server/Dockerfile", "server/requirements.txt"]
